+++ src/main/java/TheatreTicketBookingSystem/domain/Intefaces/Seating_Class.java
@@ -7,5 +7,5 @@ public interface Seating_Class {
 
     public String getClass_name();
     public int getPrice();
-    public int getSeat_id();
+    public Long getSeat_id();
 }
